requirements.txt
absl-py==0.5.0
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.3
astropy==3.0.2
attrs==18.1.0
auto-sklearn==0.4.0
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcolz==1.2.1
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
blis==0.2.4
bokeh==0.12.16
boto==2.48.0
Bottleneck==1.2.1
cachetools==3.1.0
catboost==0.10.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
conda==4.6.14
conda-build==3.10.5
conda-verify==2.0.0
ConfigSpace==0.4.7
contextlib2==0.5.5
cryptography==2.2.2
cycler==0.10.0
cymem==2.0.2
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
dataclasses==0.6
datashape==0.5.4
deap==1.2.2
decorator==4.3.0
dill==0.2.9
distributed==1.21.8
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fastai==1.0.51
fastcache==1.0.2
fastprogress==0.1.21
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
gast==0.2.0
gevent==1.3.0
glob2==0.6
gmpy2==2.0.8
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
graphviz==0.10.1
greenlet==0.4.13
grpcio==1.15.0
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.12.1
idna==2.6
image-classifiers==0.2.0
imageio==2.3.0
imagesize==1.0.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
isoweek==1.3.3
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
joblib==0.12.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
Keras==2.2.4
Keras-Applications==1.0.7
keras-contrib==2.0.8
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
liac-arff==2.3.1
llvmlite==0.23.1
locket==0.2.0
lockfile==0.12.2
lxml==4.2.1
marisa-trie==0.7.5
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
microgear==1.2.7
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.5.6
msgpack-python==0.5.6
multipledispatch==0.5.0
murmurhash==1.0.2
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
ninja==1.9.0
nltk==3.3
nokcut==0.4
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.16.3
numpydoc==0.8.0
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
odo==0.5.1
olefile==0.45.1
opencv-python==4.0.0.21
openpyxl==2.5.3
packaging==17.1
paho-mqtt==1.2.3
pandas==0.23.0
pandas-summary==0.0.6
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
plac==0.9.6
pluggy==0.6.0
ply==3.11
preshed==2.0.1
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.5
ptyprocess==0.5.2
py==1.5.3
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pydot==1.4.1
PyDrive==1.3.1
pyflakes==1.6.0
Pygments==2.2.0
pyheatmap==0.1.12
pylint==1.8.4
pynisher==0.4.2
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyrfr==0.7.4
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
pythainlp==2.0.3
python-crfsuite==0.9.6
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
requests==2.18.4
rope==0.10.7
rsa==4.0
ruamel-yaml==0.15.35
scikit-image==0.15.0
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn-crfsuite==0.3.6
sklearn-pandas==1.8.0
smac==0.8.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
spacy==2.1.3
Sphinx==1.7.4
sphinx-rtd-theme==0.4.1
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
srsly==0.0.5
statsmodels==0.9.0
stopit==1.1.2
sympy==1.1.1
tables==3.4.3
tabulate==0.8.3
tblib==1.3.2
tensorboard==1.10.0
tensorflow-gpu==1.10.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==7.0.4
tinydb==3.13.0
toolz==0.9.0
torch==1.0.1.post2
torchtext==0.3.1
torchvision==0.2.2.post3
tornado==5.0.2
TPOT==0.9.5
tqdm==4.28.1
traitlets==4.3.2
typing==3.6.4
unicodecsv==0.14.1
update-checker==0.16
uritemplate==3.0.0
urllib3==1.22
wasabi==0.2.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
xgboost==0.7.post3
xlrd==1.1.0
XlsxWriter==1.0.4
xlwt==1.3.0
zict==0.1.3
